// src/types/contentful.d.ts
export interface ICtaBanner extends Entry<ICtaBannerFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "ctaBanner";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IHeadlineFields {
  /** Internal name */
  internalName?: string | undefined;

  /** Eyebrow */
  eyebrow?: string | undefined;

  /** Title */
  title: string;

  /** In Page Nav Title */
  inPageNavTitle?: string | undefined;

  /** Section ID */
  sectionId?: string | undefined;

  /** Subhead */
  subhead?: string | undefined;

  /** Body */
  body?: Document | undefined;

  /** Alignment */
  alignment?: "left" | "center" | undefined;

  /** Call to Action */
  cta?: ILink[] | undefined;
}

export interface IHeadline extends Entry<IHeadlineFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "headline";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPerson extends Entry<IPersonFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "person";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPostFields {
  /** Theme */
  theme?: "light" | "dark" | undefined;

  /** Title */
  title: string;

  /** Type */
  type?: "blog" | undefined;

  /** Category */
  category?: "Design" | "Engineering" | "Work" | undefined;

  /** Slug */
  slug: string;

  /** Eyebrow */
  eyebrow?: string | undefined;

  /** Subhead */
  subhead?: string | undefined;

  /** Subhead Description */
  subheadDescription?: string | undefined;

  /** Text */
  text: Document;

  /** Thumbnail Text */
  thumbnailText?: string | undefined;

  /** Author */
  author?: IPerson | undefined;

  /** Published On */
  publishedOn?: string | undefined;

  /** Hero Image */
  heroImage?: Asset | undefined;

  /** Thumbnail Image */
  thumbnailImage?: Asset | undefined;

  /** ctaBanner */
  ctaBanner?: ICtaBanner[] | undefined;

  /** Metadata */
  metadata?: ISeo | undefined;

  /** Show Table of Contents */
  toc?: boolean | undefined;

  /** Display this translation? */
  translated?: boolean | undefined;
}

export interface IPost extends Entry<IPostFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "post";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPostListFields {
  /** Internal Name */
  name?: string | undefined;

  /** Theme */
  theme?: "dark" | "light" | undefined;

  /** Headline */
  headline?: IHeadline | undefined;

  /** Section ID */
  sectionId?: string | undefined;

  /** Sort */
  sort?: "title" | "publishedOn" | undefined;

  /** Order */
  order?: "asc" | "desc" | undefined;

  /** Limit */
  limit?: number | undefined;
}
// ...
export interface ISeo extends Entry<ISeoFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "seo";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}